#28 refactor: terms 패키지 이동
